update stock repository/service

+++ PruebaAA/PruebaAA.Infrastructure/Repositories/StockRepository.cs
@@ -4,6 +4,7 @@ using PruebaAA.Core.Interfaces.Repositories;
 using PruebaAA.Infrastructure.Data;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -11,7 +12,7 @@ namespace PruebaAA.Infrastructure.Repositories
 {
     public class StockRepository : IStockRepository
     {
-        private readonly PruebaAAContext _context;
+        private PruebaAAContext _context;
         public StockRepository()
         {
             _context = new PruebaAAContext();
@@ -19,8 +20,12 @@ namespace PruebaAA.Infrastructure.Repositories
         public async Task DeleteBulk()
         {
             var stocks = await GetList();
-            _context.Stocks.RemoveRange(stocks);
-            await _context.SaveChangesAsync();
+
+            if(stocks.Count() > 0)
+            {
+                _context.Stocks.RemoveRange(stocks);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task<IEnumerable<StockEntity>> GetList()
@@ -31,12 +36,45 @@ namespace PruebaAA.Infrastructure.Repositories
 
         public async Task InsertBulk(IEnumerable<StockEntity> stocks)
         {
-            foreach(StockEntity stock in stocks)
-            {
-                _context.Stocks.Add(stock);
-            }
+            int recordCount = 50000; // Realice varias pruebas y 50000 me parecio una buena opcion para la insercion de records
+            int skipCount = 0;
+            int batchCount = 0;
 
-            await _context.SaveChangesAsync();
+            //var stopwatch = new Stopwatch();
+            
+            while (skipCount <= stocks.Count())
+            {
+                //stopwatch.Start();
+
+                // Insertamos por grupos para mejorar el rendimiento y no crashear la app
+                var stockList = stocks.Skip(skipCount)
+                                .Take(recordCount)
+                                .Select(s => 
+                                new StockEntity { 
+                                    Id = s.Id, 
+                                    Product = s.Product, 
+                                    Date = s.Date, 
+                                    PointOfSale = s.PointOfSale, 
+                                    Stock = s.Stock 
+                                });
+
+                await _context.Stocks.AddRangeAsync(stockList);
+                await _context.SaveChangesAsync();
+                skipCount += recordCount;
+                
+                //stopwatch.Stop();
+                //Console.WriteLine("Tiempo de ejecucion: " + stopwatch.Elapsed.TotalSeconds);
+                //stopwatch.Restart();
+                batchCount++;
+                                
+                // Dispose el context para limpiar las miles de entidades atadas a el
+                if (batchCount >= 10)
+                {
+                    _context.Dispose();
+                    _context = new PruebaAAContext();
+                    batchCount = 0;
+                }
+            }            
         }
     }
 }
